--- src/config/constants.ts
@@ -1,13 +1,0 @@
-import 'dotenv/config'
-
-// APP constants
-export const PORT = process.env.PORT
-
-// DB Constants
-export const DB_HOST = process.env.DB_HOST
-export const DB_USER = process.env.DB_USER
-export const DB_NAME = process.env.DB_NAME
-export const DB_PASSWORD = process.env.DB_PASSWORD
-
-// JWT
-export const SECRET_KEY = process.env.SECRET_KEY
