--- src/config/database.ts
@@ -1,14 +1,0 @@
-import { Options } from 'sequelize/types'
-import { DB_HOST, DB_NAME, DB_PASSWORD, DB_USER } from './constants'
-
-export const dbConfig: Options = {
-  dialect: 'postgres',
-  host: DB_HOST,
-  username: DB_USER,
-  password: DB_PASSWORD,
-  database: DB_NAME,
-  define: {
-    timestamps: true, // gera o campo createdAt e updatedAt que contem data e hora
-    underscored: true // define o nome das tabelas e nome das colunas como snake case
-  }
-}
